# Optimize steim 1 and 2 decoders