--- a/ChangeLog
+++ b/ChangeLog
@@ -1,3 +1,8 @@
+2023.178:
+	- Optimize Steim 1 & 2 decoders.
+	- Change test for environment variable DECODE_DEBUG to compile-time define
+	- Change test for environment variable ENCODE_DEBUG to compile-time define
+
 2023.177:
 	- Incorporate ms_gswapX() functions into libmseed.h to promote inlining.