… gswap.c