Approach taken:

* Avoid potential pointer dereferencing
* Remove unnecessary values, write directly to output array
* Simplify loops, allow compiler to unroll or vectorize as it sees fit
* Remove run-time check for debugging output, now compile-time options (same for packing)