@chad-earthscope 2023-06-30
Test program for steim levels 1 and 2 decoding:
[decode_rodeo.c.zip](https://github.com/EarthScope/libmseed/files/11918886/decode_rodeo.c.zip)

The program runs the decoders in a timed loop with variable iterations.