@chad-earthscope 2023-06-30
Test OS: macOS 13.4.1
CPU: 2.4 GHz Core i9
compiler: Apple clang version 14.0.3
test program command: `./decode_rodeo -i 1000000`

| Decoder | Optimization | Original (6f5add8e3b55009d21a05abb6154e9ef324f1252) | PR #90 | This PR (e0b21aa8f51ecd7f1bc6dd7929cddcc1d036258a) |
| --- | --- | --- | --- | --- |
| Steim-2 | -O0 | 8.4 seconds | 7.9 s | 7.2 s |
| Steim-1 | -O0 | 7.0 seconds | 7.7 s | 6.5 s |
| Steim-2 | -O2 | 1.7 seconds | 1.98 s | 1.1 s |
| Steim-1 | -O3 | 1.8 seconds | 1.57 s| 0.97 s|

@Cuda-Chen Above includes test results for your PR.  I would be grateful if you could confirm the results.