@Cuda-Chen 2023-07-01
Hi @chad-earthscope ,
below table shows the results of the execution time comparison on my computer:

| Decoder | Optimization | Original (6f5add8) | This PR (e0b21aa) |
| -- | -- | -- | -- |
| Steim-2 | `-O0` | 7.50 seconds | 7.65 seconds |
| Steim-1 | `-O0` | 5.86 seconds | 5.82 seconds |
| Steim-2 | `-O2` | 2.24 seconds | 2.23 seconds |
| Steim-1 | `-O2` | 2.46 seconds | 1.76 seconds |
| Steim-2 | `-O3` | 2.79 seconds | 2.16 seconds |
| Steim-1 | `-O3` | 1.91 seconds |1.68 seconds |

## Test Specification
- CPU: Intel(R) Core(TM) i5-3230M CPU @ 2.60GHz
- OS: Linux 5.15.0-76
- compiler: GCC 11.3.0

## Test Procedures
1. Unzip `decode_rodeo.c.zip` in repo's root directory.
2. Add `decode_rodeo.c` in `LIB_SRCS` located in Makefile in repo's root directory.
3. Run `make`.
4. Run `gcc -o decode_rodeo *.o` to create `decode_rodeo` executable.
5. Run `./decode_rodeo` with the following procedures:
  -- Run `./decode_rodeo -i 1000000` three times. Each time, record the execution time of Steim-1 and Steim-2 decoding then count the average execution, respectively.
  -- Change the optimization flag in Makefile located in repo's root directory.
 6. Repeat 3. to 5.