@chad-earthscope 2023-07-02
Thanks @Cuda-Chen.  Not as much general improvement as I'd hoped, I'm surprised at very little improvement at `Steim-2 @ -O2`.

I think I'll go with this change unless you have time soon to work on it further.