@chad-earthscope 2023-07-03
More test results for different platforms.  Same procedure as used by @Cuda-Chen, average of 3 runs for each measurement.

linux 5.15.0-76
Intel(R) Xeon(R) CPU E3-1220 V2 @ 3.10GHz
gcc version 11.3.0 

| encoding | optimization | develop | #102  |
| -------- | ------------ | ------- | ----- |
| S2       |  -O0         | 7.76s   | 6.86s |
| S1       |  -O0         | 5.78s   | 5.3s |
| S2       |  -O2         | 2.06s   | 1.95s |
| S1       |  -O2         | 2.25s   | 1.55s |

linux 5.15.0-76
Intel(R) Celeron(R) CPU J3455 @ 1.50GHz
gcc version 11.3.0

| encoding | optimization | develop | #102  |
| -------- | ------------ | ------- | ----- |
| S2       |  -O0         | 15.8s   | 13.8s |
| S1       |  -O0         | 12.34s  | 10.8s |
| S2       |  -O2         | 5.0s    | 4.35s |
| S1       |  -O2         | 4.5s    | 3.5s  |

linux 5.10.103-v7
armv7l / Cortex-A53
gcc version 10.2.1 20210110 (Raspbian 10.2.1-6+rpi1)

| encoding | optimization | develop | #102  |
| -------- | ------------ | ------- | ----- |
| S2       |  -O0         | 63.5s   | 54.5s |
| S1       |  -O0         | 50.5s   | 45.4s |
| S2       |  -O2         | 16.5s   | 13.2s |
| S1       |  -O2         | 13.9s   | 11.9s |



Not a drastic improvement, but varying between 5-30% reduction in run times, and a consistent improvement in each case.

Given all of the above results, I'm going to merge this for a stable release.  We can continue to work on improvements in another PR.